Added GLMConv conversion constructor for keg types

+++ include/voxel/IntervalTree.h
@@ -135,7 +135,7 @@ public:
         if (m_tree[m_root].parent != -1) {
             return false;
         }
-        for (int i = 0; i < size(); i++) {
+        for (int i = 0; i < (int)size(); i++) {
             if (m_tree[i].parent == i) {
                 return false;
             }
